requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
psycopg2-binary==2.9.6
Flask-Mail==0.9.1
Werkzeug==2.3.4
google-api-python-client==2.88.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib
Flask-Mail
